Enable the RTC plugin
This step should cause loading of the editor to fail with an error:
"Failed to load RTC configuration settings. More information may be
available in the console"
which is fixed in the following steps.

Otherwise if the editor seems to load then check the console for:
"Failed to load plugin: rtc from url plugins/rtc/plugin.min.js"

In which case you should check that you are loading the latest version
from the cloud as older versions of TinyMCE do not bundle the plugin.