Specify a stub implementation for rtc_token_provider
This stub uses the login token to identify the user.

The reason it is not ideal is that it allows the user to access all
documents associated with the API key. A later step will restrict this
access.

Note that one more setting is still required for RTC to load.