Specify a stub implementation for rtc_encryption_provider
This stub uses a fixed password to encrypt all documents. For a keyHint
it uses the Unix epoch in ISO-8601 format.

The reason it is insecure is that an encryption key should be impossible
to guess and it should only cover one document. A later step will show
how to fix this implementation.

This is the last of the required settings so RTC will load.